requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
astunparse==1.6.3
boto3==1.42.2
botocore==1.42.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.123.8
flatbuffers==25.9.23
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
idna==3.11
jmespath==1.0.1
joblib==1.5.2
keras==3.12.0
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.3.5
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rich==14.2.0
s3transfer==0.16.0
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
starlette==0.50.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
Werkzeug==3.1.4
wrapt==2.0.1

numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0
torch==2.0.1
matplotlib==3.7.2
reportlab==4.0.4
seaborn>=0.12.0

# For diagnostics (VIF, etc.)
statsmodels>=0.14.0

# For Markowitz Optimization
cvxpy==1.3.2
PyPortfolioOpt==1.5.5

# For Deep Learning
tensorflow==2.13.0  # hoặc pytorch tùy model bạn dùng
keras==2.13.1

# Utils
python-dotenv==1.0.0
"""